+++ onsenui/CHANGELOG.md
@@ -34,6 +34,7 @@ dev
  * ons-splitter-side: Added `swipeable` property.
  * ons-list-item: Added `expandable` attribute and property. ([#2896](https://github.com/OnsenUI/OnsenUI/issues/2896)).
  * ons-list-item: Added `expand` event that fires when an expandable list item is clicked and expands or contracts.
+ * ons-button: Added `ripple` property to mirror existing `ripple` attribute.
 
  ### Misc